+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Version History
 
+## 0.0.2.1
+
+* Add MoveMethod.rewrite for full rewrite table with new data
+* Add query_part function
+* Change filter_by initialization list to string
+* Fix Clickhouse MoveMethod.delete
+* Improve execute custom query & MoveMethod operations
+* Update depends native-dumper==0.3.3.1
+* Update depends pgpack-dumper==0.3.3.1
+
 ## 0.0.2.0
 
 * Update depends native-dumper==0.3.3.0